+++ CHANGELOG.md
@@ -11,8 +11,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Removed
 
 ## [Unreleased]
+
+## [100.0.3] - 2023-04-19
 ### Fixed
 - fix for statistic filters
+- fix for counting not selected in multiselect field
+- fix for boolean field, count not selected as "No"
 
 ## [100.0.2] - 2023-04-19
 ### Changed